Backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2
django-auto-logout==0.5.1
django-cities==0.6.2
django-cors-headers==4.7.0
django-countries==7.6.1
django-model-utils==5.0.0
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
django_db==0.0.7
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
